+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Send managed trace if native trace is empty and update frame.type and error.type setting [#960](https://github.com/bugsnag/bugsnag-unity/pull/960)
 
+### Bug Fixes
+
+- Correctly handle mixed C#/Java stack traces on Android [#962](https://github.com/bugsnag/bugsnag-unity/pull/962)
+
 ## 8.8.2 (2025-12-08)
 
 ### Bug Fixes